Move 'ts/' to 'src/'.
